Router系のプラグインを入れたのでcopy系を元に戻す
