+++ react-playground/src/App.css
@@ -1,43 +1,38 @@
 .App {
-  max-width: 1024px;
-  margin: 0 auto;
-  padding: 40px 0;
   text-align: center;
 }
 
-.button {
+.App-logo {
+  height: 40vmin;
+  pointer-events: none;
+}
+
+@media (prefers-reduced-motion: no-preference) {
+  .App-logo {
+    animation: App-logo-spin infinite 20s linear;
+  }
+}
+
+.App-header {
+  background-color: #282c34;
+  min-height: 100vh;
   display: flex;
+  flex-direction: column;
   align-items: center;
-  width: 100%;
-  max-width: 300px;
-  margin: 0 auto;
-  padding: 8px 24px;
-  border: 1px solid #aaa;
-  border-radius: 10px;
-  background-color: #ffe4c4;
-  cursor: pointer;
+  justify-content: center;
+  font-size: calc(10px + 2vmin);
+  color: white;
 }
 
-.button--red {
-  background-color: #ff5757;
+.App-link {
+  color: #61dafb;
 }
 
-.button--blue {
-  background-color: #5792ff;
-}
-
-.text {
-  width: 100%;
-}
-
-.img {
-  width: 24px;
-}
-
-.title {
-  margin: 0;
-}
-
-.contents {
-  margin-top: 40px;
+@keyframes App-logo-spin {
+  from {
+    transform: rotate(0deg);
+  }
+  to {
+    transform: rotate(360deg);
+  }
 }
