+++ react-playground/src/Home.css
@@ -1,0 +1,43 @@
+.App {
+  max-width: 1024px;
+  margin: 0 auto;
+  padding: 40px 0;
+  text-align: center;
+}
+
+.button {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  max-width: 300px;
+  margin: 0 auto;
+  padding: 8px 24px;
+  border: 1px solid #aaa;
+  border-radius: 10px;
+  background-color: #ffe4c4;
+  cursor: pointer;
+}
+
+.button--red {
+  background-color: #ff5757;
+}
+
+.button--blue {
+  background-color: #5792ff;
+}
+
+.text {
+  width: 100%;
+}
+
+.img {
+  width: 24px;
+}
+
+.title {
+  margin: 0;
+}
+
+.contents {
+  margin-top: 40px;
+}
